Move Spree::Order#apply_shipping_promotions to legacy promo gem
With a properly configured promotion system, we don't need a special
"shipping" promotion handler. This moves the shipping-specific
promotion handling into the solidus_legacy_promotions gem.
